Start of the coding process
I just made the title and the background to that title back even though we will later change it to a slideshow like we talked about. I started trying to make the navigation bar but ran into trouble with visual studio code and trying to turn the items in the navi bar into links. - Andrew